DUNE_polaris_cursor/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart
pysqlite3

# HTTP Client
requests>=2.32.3

# Machine Learning & NLP
sentence-transformers>=3.0.0
torch>=2.0.0
torchvision
torchaudio
transformers>=4.30.0
spacy>=3.7.0

# Vector Database
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available

# Document Processing
pdfplumber==0.9.0
python-pptx==0.6.21
beautifulsoup4==4.12.2
pytesseract
pillow
python-docx
openpyxl
rank-bm25==0.2.2

# CLI
click==8.1.7

# Authentication (OAuth2)
authlib==1.2.1

# Environment Management
python-dotenv==1.0.0

# Utilities
pathlib2>=2.3.7
typing-extensions>=4.5.0
itsdangerous
pydantic>=2.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

pydantic>=2.9.0,<3
pdf2image

# Web Automation
playwright==1.46.0
pyee==11.1.0
cloudscraper>=1.2.70

# Embeddings

# Vector DB / Chroma
chromadb>=0.3.29
